chore: Move remaining components from old home screen folders
