+++ src/app/Components/GenericArtistSeriesRail.tsx
@@ -1,7 +1,7 @@
-import { Spacer, useColor, Text } from "@artsy/palette-mobile"
+import { Spacer, Text, useColor } from "@artsy/palette-mobile"
 import { CollectionArtistSeriesRail_collectionGroup$data } from "__generated__/CollectionArtistSeriesRail_collectionGroup.graphql"
-import { CardRailCard } from "app/Components/Home/CardRailCard"
-import { CardRailFlatList } from "app/Components/Home/CardRailFlatList"
+import { CardRailCard } from "app/Components/CardRail/CardRailCard"
+import { CardRailFlatList } from "app/Components/CardRail/CardRailFlatList"
 import { navigate } from "app/system/navigation/navigate"
 import { extractNodes } from "app/utils/extractNodes"
 import { Schema } from "app/utils/track"
