+++ src/app/Scenes/Fair/Components/FairCollections.tsx
@@ -1,9 +1,9 @@
 import { ActionType, ContextModule, OwnerType, TappedCollectionGroup } from "@artsy/cohesion"
 import { Box, BoxProps, Text, TouchableWithScale } from "@artsy/palette-mobile"
 import { FairCollections_fair$data } from "__generated__/FairCollections_fair.graphql"
+import { CARD_WIDTH } from "app/Components/CardRail/CardRailCard"
+import { CardRailFlatList } from "app/Components/CardRail/CardRailFlatList"
 import { SmallCard } from "app/Components/Cards"
-import { CARD_WIDTH } from "app/Components/Home/CardRailCard"
-import { CardRailFlatList } from "app/Components/Home/CardRailFlatList"
 import { navigate } from "app/system/navigation/navigate"
 import { compact } from "lodash"
 import { createFragmentContainer, graphql } from "react-relay"
