+++ src/app/Scenes/HomeView/Components/ActivityRail.tsx
@@ -2,7 +2,7 @@ import { Flex, Spacer, Text, useTheme } from "@artsy/palette-mobile"
 import { ActivityRail_viewer$key } from "__generated__/ActivityRail_viewer.graphql"
 import { SectionTitle } from "app/Components/SectionTitle"
 import { shouldDisplayNotification } from "app/Scenes/Activity/utils/shouldDisplayNotification"
-import { ActivityRailItem } from "app/Scenes/Home/Components/ActivityRailItem"
+import { ActivityRailItem } from "app/Scenes/HomeView/Components/ActivityRailItem"
 import HomeAnalytics from "app/Scenes/HomeView/helpers/homeAnalytics"
 import { matchRoute } from "app/routes"
 import { navigate } from "app/system/navigation/navigate"
