+++ src/app/Scenes/HomeView/Components/ActivityRailItem.tsx
