+++ src/app/Scenes/HomeView/Components/ArtistRails/ArtistCard.tsx
