+++ src/app/Scenes/HomeView/Components/ArtistRails/ArtistRail.tsx
@@ -6,11 +6,11 @@ import { Flex } from "@artsy/palette-mobile"
 import { ArtistCard_artist$data } from "__generated__/ArtistCard_artist.graphql"
 import { ArtistRailFollowMutation } from "__generated__/ArtistRailFollowMutation.graphql"
 import { ArtistRail_rail$data } from "__generated__/ArtistRail_rail.graphql"
-import { CARD_WIDTH } from "app/Components/Home/CardRailCard"
-import { CardRailFlatList, INTER_CARD_PADDING } from "app/Components/Home/CardRailFlatList"
+import { CARD_WIDTH } from "app/Components/CardRail/CardRailCard"
+import { CardRailFlatList, INTER_CARD_PADDING } from "app/Components/CardRail/CardRailFlatList"
 import { SectionTitle } from "app/Components/SectionTitle"
 import { defaultArtistVariables } from "app/Scenes/Artist/Artist"
-import { RailScrollProps } from "app/Scenes/Home/Components/types"
+import { RailScrollProps } from "app/Scenes/HomeView/Components/types"
 import HomeAnalytics from "app/Scenes/HomeView/helpers/homeAnalytics"
 import { Schema } from "app/utils/track"
 import React, { useImperativeHandle, useRef } from "react"
