+++ src/app/Scenes/HomeView/Components/ArtistRails/RecommendedArtistsRail.tsx
@@ -3,10 +3,10 @@ import { Flex, Spacer, Spinner } from "@artsy/palette-mobile"
 import { ArtistCard_artist$data } from "__generated__/ArtistCard_artist.graphql"
 import { RecommendedArtistsRailFollowMutation } from "__generated__/RecommendedArtistsRailFollowMutation.graphql"
 import { RecommendedArtistsRail_me$data } from "__generated__/RecommendedArtistsRail_me.graphql"
-import { CardRailFlatList } from "app/Components/Home/CardRailFlatList"
+import { CardRailFlatList } from "app/Components/CardRail/CardRailFlatList"
 import { SectionTitle } from "app/Components/SectionTitle"
 import { defaultArtistVariables } from "app/Scenes/Artist/Artist"
-import { RailScrollProps } from "app/Scenes/Home/Components/types"
+import { RailScrollProps } from "app/Scenes/HomeView/Components/types"
 import HomeAnalytics from "app/Scenes/HomeView/helpers/homeAnalytics"
 import { getRelayEnvironment } from "app/system/relay/defaultEnvironment"
 import { extractNodes } from "app/utils/extractNodes"
