+++ src/app/Scenes/HomeView/Components/EmailConfirmationBanner.tests.tsx
