+++ src/app/Scenes/HomeView/Components/EmailConfirmationBanner.tsx
