+++ src/app/Scenes/HomeView/Sections/HomeViewSectionActivity.tsx
@@ -4,12 +4,12 @@ import { HomeViewSectionActivityQuery } from "__generated__/HomeViewSectionActiv
 import { HomeViewSectionActivity_section$key } from "__generated__/HomeViewSectionActivity_section.graphql"
 import { SectionTitle } from "app/Components/SectionTitle"
 import { shouldDisplayNotification } from "app/Scenes/Activity/utils/shouldDisplayNotification"
-import { SeeAllCard } from "app/Scenes/Home/Components/ActivityRail"
+import { SeeAllCard } from "app/Scenes/HomeView/Components/ActivityRail"
 import {
   ACTIVITY_RAIL_ARTWORK_IMAGE_SIZE,
   ACTIVITY_RAIL_ITEM_WIDTH,
   ActivityRailItem,
-} from "app/Scenes/Home/Components/ActivityRailItem"
+} from "app/Scenes/HomeView/Components/ActivityRailItem"
 import { HomeViewSectionSentinel } from "app/Scenes/HomeView/Components/HomeViewSectionSentinel"
 import { SectionSharedProps } from "app/Scenes/HomeView/Sections/Section"
 import {
