+++ src/app/Scenes/MyCollection/Components/MyCollectionZeroState.tsx
@@ -1,7 +1,7 @@
 import { ActionType, AddCollectedArtwork, ContextModule, OwnerType } from "@artsy/cohesion"
-import { useSpace, Flex, Button, Text, Tabs, Box } from "@artsy/palette-mobile"
+import { Box, Button, Flex, Tabs, Text, useSpace } from "@artsy/palette-mobile"
 import { ZeroState } from "app/Components/States/ZeroState"
-import { ModalCarousel } from "app/Scenes/Home/Components/ModalCarouselComponents/ModalCarousel"
+import { ModalCarousel } from "app/Scenes/HomeView/Components/ModalCarouselComponents/ModalCarousel"
 import { MyCollectionTabsStore } from "app/Scenes/MyCollection/State/MyCollectionTabsStore"
 import { debounce } from "lodash"
 import { useState } from "react"
