+++ src/app/Scenes/Search/CuratedCollectionItem.tsx
@@ -2,7 +2,7 @@ import { ActionType, ContextModule, OwnerType } from "@artsy/cohesion"
 import { TappedCollectionGroup } from "@artsy/cohesion/dist/Schema/Events/Tap"
 import { Box, Text, TextProps } from "@artsy/palette-mobile"
 import { CuratedCollectionItem_collection$key } from "__generated__/CuratedCollectionItem_collection.graphql"
-import { CardRailCard, CardRailMetadataContainer } from "app/Components/Home/CardRailCard"
+import { CardRailCard, CardRailMetadataContainer } from "app/Components/CardRail/CardRailCard"
 import { ThreeUpImageLayout } from "app/Components/ThreeUpImageLayout"
 import { navigate } from "app/system/navigation/navigate"
 import { extractNodes } from "app/utils/extractNodes"
