+++ src/app/Scenes/Search/TrendingArtists.tsx
@@ -1,14 +1,14 @@
 import { ActionType, ContextModule, OwnerType, TappedArtistGroup } from "@artsy/cohesion"
-import { Spacer, Flex, Box, BoxProps, Spinner } from "@artsy/palette-mobile"
+import { Box, BoxProps, Flex, Spacer, Spinner } from "@artsy/palette-mobile"
 import { SearchQuery } from "__generated__/SearchQuery.graphql"
 import { TrendingArtists_query$key } from "__generated__/TrendingArtists_query.graphql"
-import { ArtistCardContainer as ArtistCard } from "app/Components/Home/ArtistRails/ArtistCard"
-import { CardRailFlatList } from "app/Components/Home/CardRailFlatList"
+import { CardRailFlatList } from "app/Components/CardRail/CardRailFlatList"
 import { SectionTitle } from "app/Components/SectionTitle"
+import { ArtistCardContainer as ArtistCard } from "app/Scenes/HomeView/Components/ArtistRails/ArtistCard"
 import { navigate } from "app/system/navigation/navigate"
 import { extractNodes } from "app/utils/extractNodes"
 import { isTablet } from "react-native-device-info"
-import { usePaginationFragment, graphql } from "react-relay"
+import { graphql, usePaginationFragment } from "react-relay"
 import { useTracking } from "react-tracking"
 import { TrendingArtistCard } from "./components/TrendingArtistCard"
