+++ src/app/Scenes/ViewingRoom/Components/ViewingRoomsListFeatured.tsx
@@ -3,7 +3,7 @@ import { Spacer, Touchable } from "@artsy/palette-mobile"
 import { ViewingRoomsListFeatured_featured$key } from "__generated__/ViewingRoomsListFeatured_featured.graphql"
 import { AboveTheFoldFlatList } from "app/Components/AboveTheFoldFlatList"
 import { MediumCard } from "app/Components/Cards"
-import { RailScrollProps } from "app/Scenes/Home/Components/types"
+import { RailScrollProps } from "app/Scenes/HomeView/Components/types"
 import { navigate } from "app/system/navigation/navigate"
 import { extractNodes } from "app/utils/extractNodes"
 import { Schema } from "app/utils/track"
