+++ src/app/Scenes/ViewingRoom/ViewingRoomsList.tsx
@@ -1,11 +1,11 @@
-import { Spacer, Flex, useSpace } from "@artsy/palette-mobile"
+import { Flex, Spacer, useSpace } from "@artsy/palette-mobile"
 import { ViewingRoomsListFeatured_featured$key } from "__generated__/ViewingRoomsListFeatured_featured.graphql"
 import { ViewingRoomsListQuery } from "__generated__/ViewingRoomsListQuery.graphql"
 import { ViewingRoomsList_viewingRooms$key } from "__generated__/ViewingRoomsList_viewingRooms.graphql"
 import { PageWithSimpleHeader } from "app/Components/PageWithSimpleHeader"
 import { SectionTitle } from "app/Components/SectionTitle"
 import { PAGE_SIZE } from "app/Components/constants"
-import { RailScrollRef } from "app/Scenes/Home/Components/types"
+import { RailScrollRef } from "app/Scenes/HomeView/Components/types"
 import { extractNodes } from "app/utils/extractNodes"
 import { useScreenDimensions } from "app/utils/hooks"
 import { PlaceholderBox, PlaceholderText, ProvidePlaceholderContext } from "app/utils/placeholders"
@@ -13,7 +13,7 @@ import { ProvideScreenTracking, Schema } from "app/utils/track"
 import { times } from "lodash"
 import { Fragment, Suspense, useRef, useState } from "react"
 import { FlatList, RefreshControl } from "react-native"
-import { useLazyLoadQuery, usePaginationFragment, graphql, useFragment } from "react-relay"
+import { graphql, useFragment, useLazyLoadQuery, usePaginationFragment } from "react-relay"
 import { featuredFragment, FeaturedRail } from "./Components/ViewingRoomsListFeatured"
 import { ViewingRoomsListItem } from "./Components/ViewingRoomsListItem"
